+++ challenges/web/scratchpad/README.md
@@ -1,0 +1,7 @@
+# Scratchpad
+
+A web CTF challenge with a scratchpad app that allows users to write quick notes but with the convenience of version tracking.
+
+## Deployment
+
+Run src/deploy-docker.sh and browser to localhost:1337
