+++ challenges/web/scratchpad/src/config/nginx.conf
@@ -1,0 +1,31 @@
+worker_processes 1;
+
+events {
+    worker_connections 1024;
+}
+
+http {
+    include       mime.types;
+    default_type  application/octet-stream;
+    sendfile        on;
+    keepalive_timeout  65;
+
+    server {
+        listen 80;
+        server_name localhost;
+
+        location /api/ {
+            proxy_pass http://localhost:8080/api/;
+            proxy_http_version 1.1;
+            proxy_set_header Upgrade $http_upgrade;
+            proxy_set_header Connection 'upgrade';
+            proxy_set_header Host $host;
+        }
+
+        location / {
+            root /app/html;
+            index index.html;
+            try_files $uri /index.html;
+        }
+    }
+}
