+++ challenges/web/scratchpad/src/flag.txt
@@ -1,0 +1,1 @@
+SVUSCG{t3st_fl4g}
