--- challenges/web/versioned/src/deploy-docker.sh
@@ -1,2 +1,0 @@
-docker build -t uscg-open-web-versioned . && \
-docker run -p 1337:80 --rm uscg-open-web-versioned
