--- challenges/web/versioned/src/frontend/src/index.css
@@ -1,8 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-body {
-  font-family: system-ui, sans-serif;
-  background-color: #f9f9f9;
-}
